Remove autocomma and autostatementline Cpppp attributes
These cannot be attributes because they would change the
validity of generated code without the template's knowledge.
Only the template will know whether it is appropriate to
disable these features.  It also probably needs to be
lexically scoped, when that happens.  I'm just removing
the attribute accessors and constructor initializers
for now.