New feature convert_linecomment_to_c89
This is a new attribute on CodeGen::Cpppp.  It get applied during
new method get_filtered_output, using the CParser to find comment
tokens starting with '//'.  It has a corresponding cmdline switch
on the cpppp tool.  I also upgraded the Parser tokens to objects,
and added missing attribute accessors to the Template objects.