Detect indent from text of template and provide that as context
For code generating modules, it makes sense to match the indent of
the surrounding template.  In order to do that, the template
engine needs to analyze the text of the template.  It's difficult
to do that in the middle of _render_code_block, so I decided to
detect it during the initial parse of the template and then provide
that in localized globals during the evaluation of the block.

This code assumes that the indent remains constant for the entire
template.  If not, a user would need to override it on the fly
using calls to $self->indent(...)