Fix Win32 implementation of format_commandline
First, the unit test was broken because it didn't consider the
quoting that would happen if the path contained unsafe characters,
such as the backslashes on Win32.  But, Unix paths with a space
would have also triggered that bug.

This led to the realization that bash escaping wasn't appropriate
anyway for someone working on a Win32 platform, so I should choose
a platform-specific implementation for this function.

I created a new ::Platform utility library, and for now, just emit
the Win32::ShellQuote::quote_native for the @ARGV all on one line.
I don't think there's a way to line-wrap a windows command in a
way that pastes correctly into the shell.

I then added Win32::ShellQuote as a conditional dependency.